Ajustes do PR e Implementado Testes

+++ src/SME.SGP.Infra/Dtos/Frequencia/FrequenciaDiariaAlunoDto.cs
@@ -6,10 +6,10 @@ namespace SME.SGP.Infra
     {
         public long? Id { get; set; }
         public DateTime DataAula { get; set; }
-        public int QuantidadeAulas { get; set; }
+        public long QuantidadeAulas { get; set; }
         public int QuantidadePresenca { get; set; }
         public int QuantidadeRemoto { get; set; }
-        public int QuantidadeAusencia { get; set; }
+        public long QuantidadeAusencia { get; set; }
         public string Motivo { get; set; }
     }
 }
